FIX Task #8: FFT (Fast Fourier Transform) using cuFFT. Владимирова Юлия. 3822Б1ПР2

+++ 3822B1PE2/8_fft_cufft/vladimirova_julia/fft_cufft.cu
@@ -15,7 +15,7 @@ std::vector<float> FffCUFFT(const std::vector<float>& input, int batch) {
     int n = i_size / (2 * batch);
     int comp_size = n * batch;
     int d_size = i_size * sizeof(float);
-    int f = 1.0 / n;
+    float f = 1.0 / n;
     std::vector<float> output(i_size);
